refactor: extract building the Learning MFE link to utils
This will allow us to reuse this function.
